The 3 main scripts
This entire simulation is on these 3 scripts. They have comments describing their functions in the scripts
The camera Controls are to control the camera movement
the UserSpawn is how the user creates the simulation, it holds all the references to the ui elements that affect the variables
the CubeGrow is the essential script on start it will iterate all the positions and roll dice to spawn a cube, set its color and reduce its rate.